chore: fix more eslint errors
