+++ apps/lockfile-explorer/.eslintrc.js
@@ -5,8 +5,6 @@ module.exports = {
   extends: ['@rushstack/eslint-config/profile/web-app', '@rushstack/eslint-config/mixins/react'],
   parserOptions: { tsconfigRootDir: __dirname },
   rules: {
-    'no-use-before-define': 'off',
-    '@typescript-eslint/no-use-before-define': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/no-floating-promises': 'off'
   }
